+++ packages/gradient-controls/README.md
@@ -1,0 +1,5 @@
+# @fabricjs/gradient-controls
+
+Linear gradient control helpers for Fabric.js.
+
+This package exposes `createLinearGradientControls` as a standalone package that depends on `@fabricjs/core`.
