+++ packages/gradient-controls/package.json
@@ -1,0 +1,36 @@
+{
+  "name": "@fabricjs/gradient-controls",
+  "description": "Linear gradient control helpers for Fabric.js.",
+  "version": "7.4.0",
+  "homepage": "http://fabricjs.com/",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/fabricjs/fabric.js.git",
+    "directory": "packages/gradient-controls"
+  },
+  "bugs": {
+    "url": "https://github.com/fabricjs/fabric.js/issues"
+  },
+  "license": "MIT",
+  "sideEffects": false,
+  "module": "./dist/index.mjs",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.mjs",
+      "default": "./dist/index.mjs"
+    },
+    "./package.json": "./package.json"
+  },
+  "files": [
+    "dist",
+    "README.md"
+  ],
+  "dependencies": {
+    "@fabricjs/core": "workspace:^"
+  },
+  "engines": {
+    "node": ">=20.0.0"
+  }
+}
