+++ packages/gradient-controls/src/index.ts
@@ -1,0 +1,1 @@
+export { createLinearGradientControls } from './linearGradientControls';
