+++ tsconfig-extensions.json
@@ -31,7 +31,8 @@
     "paths": {
       "fabric": ["./index"],
       "fabric/node": ["./index.node"],
-      "fabric/extensions": ["./extensions/index"]
+      "fabric/extensions": ["./extensions/index"],
+      "@fabricjs/core": ["./index"]
     },
     // "paths": {} /* Specify a set of entries that re-map imports to additional lookup locations. */,
     // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
