+++ tsconfig.spec.json
@@ -10,6 +10,7 @@
       "fabric/extensions": ["./extensions"],
       "@fabricjs/core": ["./fabric"],
       "@fabricjs/browser": ["./packages/browser/src/index"],
+      "@fabricjs/gradient-controls": ["./packages/gradient-controls/src/index"],
       "@fabricjs/node": ["./packages/node/src/index"]
     }
   },
